refactor: move helper function out to helpers
